+++ eventdisplayAT/ATEventDrawTask.hh
@@ -80,6 +80,7 @@ class ATEventDrawTask : public FairTask
     void SetProtoMap(TString map) {fMap = map;}
 
     void SetMultiHit(Int_t hitMax);
+    void SetAlgorithm(Int_t val) {fRANSACAlg = val;};
 
   protected :
     virtual void DrawPadPlane();
@@ -250,6 +251,7 @@ class ATEventDrawTask : public FairTask
     Bool_t fSaveTextData;
     Float_t f3DThreshold;
     Bool_t fIsRawData;
+    Int_t fRANSACAlg;
 
     TF1 *fHoughLinearFit;
     TF1 *fRansacLinearFit;
