+++ macro/Simulation/d2He/outputFiles/event.dat
@@ -1,15 +1,39 @@
- Event ID : 98
-17.1788 15.5519 16 20 2.3017
-14.7247 14.2027 46 35 14.1873
--63.8071 -201.367 834 429 1.19375
--44.1741 -138.889 856 440 2.11878
--176.696 90.85 896 460 4.29581
--201.238 99.3513 896 460 2.04034
--196.329 96.3821 896 460 2.25206
--215.962 107.853 896 460 1.64362
--211.054 104.883 896 460 1.68435
--127.614 62.3768 898 461 4.79426
--161.972 82.3487 898 461 1.6509
--171.788 87.8808 898 461 1.39776
--191.421 99.3513 898 461 2.06352
--24.5412 11.3012 900 462 1.2973
+ Event ID : 19
+-83.44 45.3065 78 51 8.85776
+-80.9859 43.9573 86 55 20.2376
+-80.9859 41.0558 98 61 4.11828
+-78.5318 39.7067 106 65 3.6839
+-76.0777 41.0558 114 69 4.77317
+-73.6235 39.7067 134 79 17.2706
+-73.6235 36.8052 144 84 2.52463
+-71.1694 35.456 154 89 3.79099
+-68.7153 36.8052 162 93 5.11423
+-66.2612 35.456 182 103 7.53939
+-63.8071 31.2053 196 110 1.63536
+-61.353 32.5545 208 116 19.6048
+-58.8988 31.2053 226 125 1.60153
+-56.4447 32.5545 236 130 1.45855
+-53.9906 31.2053 250 137 1.12682
+-51.5365 26.9547 272 148 3.92676
+-49.0824 28.3038 284 154 20.4837
+-46.6282 26.9547 296 160 2.43003
+-41.72 24.0532 326 175 1.65794
+-39.2659 22.704 344 184 3.68983
+-36.8118 24.0532 356 190 2.77417
+-34.3577 22.704 364 194 2.98022
+-31.9035 18.4534 384 204 3.62352
+-29.4494 19.8025 398 211 8.09733
+-26.9953 18.4534 414 219 5.16166
+-22.0871 15.5519 440 232 5.02083
+-19.6329 14.2027 458 241 2.32437
+-17.1788 15.5519 468 246 2.55813
+2.45411 -77.9626 684 354 1.87064
+-7.62939e-06 -87.813 692 358 1.03536
+-4.90823 -138.889 736 380 2.95834
+-4.90823 -150.359 744 384 3.61185
+-4.90824 -172.894 766 395 1.78767
+-9.81647 -181.395 770 397 4.80879
+-9.81647 -192.866 782 403 2.3072
+-14.7247 -201.367 788 406 1.00068
+-9.81648 -198.398 788 406 1.67877
+-9.81647 -209.868 796 410 1.02277
