+++ macro/Simulation/d2He/rundigi_d2He.C
@@ -1,10 +1,10 @@
 void rundigi_d2He
 (TString mcFile = "outputFiles/attpcsim_d2He.root",//attpcsim_d2He_14O_07atm_100000.root
-//(TString mcFile = "outputFiles/attpcsim_d2He_simon3.root",//attpcsim_d2He_14O_07atm_100000.root
+//(TString mcFile = "outputFiles/attpcsim_d2He_1e6pps.root",//attpcsim_d2He_14O_07atm_100000.root
 TString digiParFile = "../../../parameters/ATTPC.d2He.par",
 TString mapParFile = "../../../scripts/Lookup20150611.xml",
 TString parFile = "outputFiles/attpcpar_d2He.root")    //"attpcpar_d2He_test_10k.root")//attpcpar_d2He_14O_07atm_100000.root
-//TString parFile = "outputFiles/attpcpar_d2He_simon3.root")    //"attpcpar_d2He_test_10k.root")//attpcpar_d2He_14O_07atm_100000.root
+//TString parFile = "outputFiles/attpcpar_d2He_1e6pps.root")    //"attpcpar_d2He_test_10k.root")//attpcpar_d2He_14O_07atm_100000.root
 //TString trigParFile = "../../../parameters/AT.trigger.par")
 {
 
@@ -79,8 +79,9 @@ TString parFile = "outputFiles/attpcpar_d2He.root")    //"attpcpar_d2He_test_10k
   //RandTask ->SetModelType(1);
   //RandTask ->SetFullMode();
   RandTask->SetTiltAngle(0.0);
-  RandTask->SetDistanceThreshold(7.5);
-  RandTask->SetMinHitsLine(10);
+  RandTask->SetDistanceThreshold(15.0);
+  RandTask->SetMinHitsLine(7);
+  RandTask->SetAlgorithm(1); // 0=PCL ransac; 1=Homemade Ransac
 
 
 
@@ -108,7 +109,7 @@ TString parFile = "outputFiles/attpcpar_d2He.root")    //"attpcpar_d2He_test_10k
 
   fRun -> Init();
   //fRun -> Run(0,nEvents);
-  fRun -> Run(0,100);
+  fRun -> Run(0,50);
 
   std::cout << std::endl << std::endl;
   std::cout << "Macro finished succesfully."  << std::endl << std::endl;
