+++ reco/CMakeLists.txt
@@ -23,7 +23,6 @@ ${CMAKE_SOURCE_DIR}/reco/ATAnalysis
 ${CMAKE_SOURCE_DIR}/reco/ATRansac
 ${CMAKE_SOURCE_DIR}/reco/ATPatternRecognition
 ${CMAKE_SOURCE_DIR}/tpc
-${CMAKE_SOURCE_DIR}/reco/ATRansacMod
 
 )
 
@@ -118,7 +117,7 @@ ATAnalysis/ATTrackingAnalysis.cc
 ATAnalysis/ATd2HeAnalysis.cc
 
 ATRansac/ATRansac.cc
-ATRansacMod/ATRansacMod.cc
+ATPatternRecognition/ATRansacMod.cc
 
 ATPatternRecognition/ATPRA.cc
 ATPatternRecognition/ATTrackFinderHC.cc
